backend/requirements.txt
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==44.0.2
distlib==0.3.8
filelock==3.15.3
Flask==3.0.3
flask-restx==1.3.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
platformdirs==4.2.2
pymysql==1.1.1
python-dotenv==1.0.1
urllib3==2.2.2
virtualenv==20.26.2
Werkzeug==3.0.3
